test: split PDF visual caption lane coverage
